+++ internal/commands/server/show.go
@@ -8,6 +8,7 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/completion"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/format"
+	"github.com/UpCloudLtd/upcloud-cli/v2/internal/labels"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/resolver"
 	"github.com/UpCloudLtd/upcloud-cli/v2/internal/ui"
@@ -139,6 +140,7 @@ func (s *showCommand) Execute(exec commands.Executor, uuid string) (output.Outpu
 				},
 			},
 		},
+		labels.GetLabelsSection(server.Labels),
 		output.CombinedSection{
 			Key:   "storage",
 			Title: "Storage: (Flags: B = bootdisk, P = part of plan)",